+++ packages/stacked_tools/lib/src/commands/create/create_service_command.dart
@@ -4,11 +4,9 @@ import 'package:stacked_tools/src/constants/message_constants.dart';
 import 'package:stacked_tools/src/locator.dart';
 import 'package:stacked_tools/src/mixins/project_structure_validator_mixin.dart';
 import 'package:stacked_tools/src/services/process_service.dart';
+import 'package:stacked_tools/src/services/pubspec_service.dart';
 import 'package:stacked_tools/src/services/template_service.dart';
 import 'package:stacked_tools/src/templates/template_constants.dart';
-
-import '../../services/config_service.dart';
-import '../../services/pubspec_service.dart';
 
 class CreateServiceCommand extends Command with ProjectStructureValidator {
   final _templateService = locator<TemplateService>();
@@ -31,15 +29,13 @@ class CreateServiceCommand extends Command with ProjectStructureValidator {
     argParser.addOption(
       ksLineLength,
       abbr: 'l',
-      help: 'The length of the line that is used for formatting',
+      help: kCommandHelpLineLength,
       valueHelp: '80',
     );
   }
 
   @override
   Future<void> run() async {
-    await locator<ConfigService>().loadConfig();
-
     final outputPath = argResults!.rest.length > 1 ? argResults!.rest[1] : null;
     _processService.formattingLineLength = argResults?[ksLineLength];
     await _pubspecService.initialise(workingDirectory: outputPath);