+++ packages/stacked_tools/lib/src/commands/create/create_view_command.dart
@@ -44,10 +44,8 @@ class CreateViewCommand extends Command with ProjectStructureValidator {
 
   @override
   Future<void> run() async {
-    await _configService.loadConfig();
-
     final outputPath = argResults!.rest.length > 1 ? argResults!.rest[1] : null;
-    _processService.formattingLineLength = argResults?[ksLineLength];
+    _processService.formattingLineLength = argResults![ksLineLength];
     await _pubspecService.initialise(workingDirectory: outputPath);
     await validateStructure(outputPath: outputPath);