+++ packages/stacked_tools/lib/src/templates/compiled_templates.dart
@@ -1,7 +1,6 @@
 /// NOTE: This is generated code from the compileTemplates command. Do not modify by hand
 ///       This file should be checked into source control.
 
-
 // -------- HomeViewmodelTest Template Data ----------
 
 const String kAppTemplateHomeViewmodelTestPath =
@@ -53,11 +52,9 @@ void main() {
 
 // --------------------------------------------------
 
-
 // -------- TestHelpers Template Data ----------
 
-const String kAppTemplateTestHelpersPath =
-    'test/helpers/test_helpers.dart.stk';
+const String kAppTemplateTestHelpersPath = 'test/helpers/test_helpers.dart.stk';
 
 const String kAppTemplateTestHelpersContent = '''
 import 'package:mockito/annotations.dart';
@@ -143,11 +140,9 @@ void _removeRegistrationIfExists<T extends Object>() {
 
 // --------------------------------------------------
 
-
 // -------- StackedConfigJsonStk Template Data ----------
 
-const String kAppTemplateStackedConfigJsonStkPath =
-    'stacked.config.json.stk';
+const String kAppTemplateStackedConfigJsonStkPath = 'stacked.config.json.stk';
 
 const String kAppTemplateStackedConfigJsonStkContent = '''
 {
@@ -159,17 +154,16 @@ const String kAppTemplateStackedConfigJsonStkContent = '''
     "test_views_path": "test/viewmodels",
     "locator_name": "locator",
     "register_mocks_function": "registerServices",
-    "v1": false
+    "v1": false,
+    "line_length": 80
 }
 ''';
 
 // --------------------------------------------------
 
-
 // -------- READMEMdStk Template Data ----------
 
-const String kAppTemplateREADMEMdStkPath =
-    'README.md.stk';
+const String kAppTemplateREADMEMdStkPath = 'README.md.stk';
 
 const String kAppTemplateREADMEMdStkContent = '''
 # stacked_app
@@ -193,11 +187,9 @@ samples, guidance on mobile development, and a full API reference.
 
 // --------------------------------------------------
 
-
 // -------- Main Template Data ----------
 
-const String kAppTemplateMainPath =
-    'lib/main.dart.stk';
+const String kAppTemplateMainPath = 'lib/main.dart.stk';
 
 const String kAppTemplateMainContent = '''
 import 'package:flutter/material.dart';
@@ -242,7 +234,6 @@ class MyApp extends StatelessWidget {
 
 // --------------------------------------------------
 
-
 // -------- SetupDialogUi Template Data ----------
 
 const String kAppTemplateSetupDialogUiPath =
@@ -268,7 +259,6 @@ void setupDialogUi() {
 ''';
 
 // --------------------------------------------------
-
 
 // -------- SetupBottomSheetUi Template Data ----------
 
@@ -300,11 +290,9 @@ void setupBottomSheetUi() {
 
 // --------------------------------------------------
 
-
 // -------- UiHelpers Template Data ----------
 
-const String kAppTemplateUiHelpersPath =
-    'lib/ui/common/ui_helpers.dart.stk';
+const String kAppTemplateUiHelpersPath = 'lib/ui/common/ui_helpers.dart.stk';
 
 const String kAppTemplateUiHelpersContent = '''
 import 'dart:math';
@@ -390,11 +378,9 @@ double getResponsiveFontSize(BuildContext context,
 
 // --------------------------------------------------
 
-
 // -------- AppStrings Template Data ----------
 
-const String kAppTemplateAppStringsPath =
-    'lib/ui/common/app_strings.dart.stk';
+const String kAppTemplateAppStringsPath = 'lib/ui/common/app_strings.dart.stk';
 
 const String kAppTemplateAppStringsContent = '''
 const String ksHomeBottomSheetTitle = 'Build Great Apps!';
@@ -405,11 +391,9 @@ const String ksHomeBottomSheetDescription =
 
 // --------------------------------------------------
 
-
 // -------- AppColors Template Data ----------
 
-const String kAppTemplateAppColorsPath =
-    'lib/ui/common/app_colors.dart.stk';
+const String kAppTemplateAppColorsPath = 'lib/ui/common/app_colors.dart.stk';
 
 const String kAppTemplateAppColorsContent = '''
 import 'package:flutter/material.dart';
@@ -425,7 +409,6 @@ const Color kcBackgroundColor = kcDarkGreyColor;
 ''';
 
 // --------------------------------------------------
-
 
 // -------- NoticeSheet Template Data ----------
 
@@ -484,7 +467,6 @@ class NoticeSheet extends StatelessWidget {
 
 // --------------------------------------------------
 
-
 // -------- NoticeSheetViewmodel Template Data ----------
 
 const String kAppTemplateNoticeSheetViewmodelPath =
@@ -495,7 +477,6 @@ const String kAppTemplateNoticeSheetViewmodelContent = '''
 ''';
 
 // --------------------------------------------------
-
 
 // -------- InfoAlertDialog Template Data ----------
 
@@ -605,7 +586,6 @@ class InfoAlertDialog extends StatelessWidget {
 
 // --------------------------------------------------
 
-
 // -------- InfoAlertViewmodel Template Data ----------
 
 const String kAppTemplateInfoAlertViewmodelPath =
@@ -618,7 +598,6 @@ const String kAppTemplateInfoAlertViewmodelContent = '''
 ''';
 
 // --------------------------------------------------
-
 
 // -------- HomeViewV1 Template Data ----------
 
@@ -708,11 +687,9 @@ class HomeView extends StatelessWidget {
 
 // --------------------------------------------------
 
-
 // -------- HomeView Template Data ----------
 
-const String kAppTemplateHomeViewPath =
-    'lib/ui/views/home/home_view.dart.stk';
+const String kAppTemplateHomeViewPath = 'lib/ui/views/home/home_view.dart.stk';
 
 const String kAppTemplateHomeViewContent = '''
 import 'package:flutter/material.dart';
@@ -800,7 +777,6 @@ class HomeView extends StackedView<HomeViewModel> {
 
 // --------------------------------------------------
 
-
 // -------- HomeViewmodel Template Data ----------
 
 const String kAppTemplateHomeViewmodelPath =
@@ -847,7 +823,6 @@ class HomeViewModel extends BaseViewModel {
 
 // --------------------------------------------------
 
-
 // -------- StartupViewmodel Template Data ----------
 
 const String kAppTemplateStartupViewmodelPath =
@@ -876,7 +851,6 @@ class StartupViewModel extends BaseViewModel {
 ''';
 
 // --------------------------------------------------
-
 
 // -------- StartupViewV1 Template Data ----------
 
@@ -942,7 +916,6 @@ class StartupView extends StatelessWidget {
 ''';
 
 // --------------------------------------------------
-
 
 // -------- StartupView Template Data ----------
 
@@ -1014,11 +987,9 @@ class StartupView extends StackedView<StartupViewModel> {
 
 // --------------------------------------------------
 
-
 // -------- App Template Data ----------
 
-const String kAppTemplateAppPath =
-    'lib/app/app.dart.stk';
+const String kAppTemplateAppPath = 'lib/app/app.dart.stk';
 
 const String kAppTemplateAppContent = '''
 import 'package:stacked/stacked_annotations.dart';
@@ -1043,11 +1014,9 @@ class App {}
 
 // --------------------------------------------------
 
-
 // -------- DialogType Template Data ----------
 
-const String kAppTemplateDialogTypePath =
-    'lib/enums/dialog_type.dart.stk';
+const String kAppTemplateDialogTypePath = 'lib/enums/dialog_type.dart.stk';
 
 const String kAppTemplateDialogTypeContent = '''
 enum DialogType {
@@ -1057,7 +1026,6 @@ enum DialogType {
 ''';
 
 // --------------------------------------------------
-
 
 // -------- BottomSheetType Template Data ----------
 
@@ -1073,11 +1041,9 @@ enum BottomSheetType {
 
 // --------------------------------------------------
 
-
 // -------- PubspecYamlStk Template Data ----------
 
-const String kAppTemplatePubspecYamlStkPath =
-    'pubspec.yaml.stk';
+const String kAppTemplatePubspecYamlStkPath = 'pubspec.yaml.stk';
 
 const String kAppTemplatePubspecYamlStkContent = '''
 name: {{packageName}}
@@ -1185,7 +1151,6 @@ flutter:
 
 // --------------------------------------------------
 
-
 // -------- GenericViewmodelTest Template Data ----------
 
 const String kViewTemplateGenericViewmodelTestPath =
@@ -1208,7 +1173,6 @@ void main() {
 
 // --------------------------------------------------
 
-
 // -------- GenericViewmodel Template Data ----------
 
 const String kViewTemplateGenericViewmodelPath =
@@ -1222,7 +1186,6 @@ class {{viewModelName}} extends BaseViewModel {
 ''';
 
 // --------------------------------------------------
-
 
 // -------- GenericView Template Data ----------
 
@@ -1257,7 +1220,6 @@ class {{viewName}} extends StackedView<{{viewModelName}}> {
 
 // --------------------------------------------------
 
-
 // -------- GenericViewV1 Template Data ----------
 
 const String kViewTemplateGenericViewV1Path =
@@ -1289,7 +1251,6 @@ class {{viewName}} extends StatelessWidget {
 
 // --------------------------------------------------
 
-
 // -------- GenericServiceTest Template Data ----------
 
 const String kServiceTemplateGenericServiceTestPath =
@@ -1312,7 +1273,6 @@ void main() {
 
 // --------------------------------------------------
 
-
 // -------- GenericService Template Data ----------
 
 const String kServiceTemplateGenericServicePath =
@@ -1325,4 +1285,3 @@ class {{serviceName}}Service {
 ''';
 
 // --------------------------------------------------
-